user ve admin girişi/yetkileri ayarlandı

+++ OrderManagement/routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CategoryController;
 use App\Http\Controllers\ProductController;
 use Illuminate\Http\Request;
@@ -16,24 +17,32 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    return $request->user();
+Route::group([
+    'middleware' => 'api',
+    'prefix' => 'auth'
+
+], function ($router) {
+    Route::post('/login', [AuthController::class, 'login']);
+    Route::post('/register', [AuthController::class, 'register']);
+    Route::post('/logout', [AuthController::class, 'logout']);
+    Route::post('/token/refresh', [AuthController::class, 'refresh']);
+    Route::get('/user', [AuthController::class, 'user']);
 });
 
 
 Route::prefix('categories')->group(function () {
-    Route::get('/', [CategoryController::class, 'index']);
-    Route::get('{id}', [CategoryController::class, 'show']);
-    Route::post('/', [CategoryController::class, 'store']);
-    Route::put('{id}', [CategoryController::class, 'update']);
-    Route::delete('{id}', [CategoryController::class, 'destroy']);
+    Route::get('/', [CategoryController::class, 'index'])->middleware('auth');
+    Route::get('{id}', [CategoryController::class, 'show'])->middleware('auth');
+    Route::post('/', [CategoryController::class, 'store'])->middleware('is_admin');
+    Route::put('{id}', [CategoryController::class, 'update'])->middleware('is_admin');
+    Route::delete('{id}', [CategoryController::class, 'destroy'])->middleware('is_admin');
 });
 Route::prefix('products')->group(function () {
     Route::controller(ProductController::class)->group(function () {
-        Route::get('/', 'index');
-        Route::get('{id}', 'show');
-        Route::post('/', 'store');
-        Route::put('{id}', 'update');
-        Route::delete('{id}', 'destroy');
+        Route::get('/', 'index')->middleware('auth');
+        Route::get('{id}', 'show')->middleware('auth');
+        Route::post('/', 'store')->middleware('is_admin');
+        Route::put('{id}', 'update')->middleware('is_admin');
+        Route::delete('{id}', 'destroy')->middleware('is_admin');
     });
 });
